Move KontoCheckIbanValidator together with its cohesive code
