+++ contexts/PaymentContext/src/UseCases/GenerateIban/GenerateIbanUseCase.php
@@ -5,8 +5,8 @@ declare( strict_types = 1 );
 namespace WMDE\Fundraising\PaymentContext\UseCases\GenerateIban;
 
 use WMDE\Fundraising\PaymentContext\Domain\BankDataConverter;
+use WMDE\Fundraising\PaymentContext\Domain\KontoCheckIbanValidator;
 use WMDE\Fundraising\PaymentContext\ResponseModel\IbanResponse;
-use WMDE\Fundraising\Frontend\Validation\KontoCheckIbanValidator;
 
 /**
  * @licence GNU GPL v2+
